+++ firmware/matrix_math.h
@@ -1,7 +1,7 @@
 #ifndef MATRIX_HEADER
 #define MATRIX_HEADER
 
-void matrix_inverse(const long double m[16], long double invOut[16]);
-void mat_vec_dot(const long double m[16], long double vec[4], long double vec_out[4]);
+void matrix_inverse(const double m[16], double invOut[16]);
+void mat_vec_dot(const double m[16], double vec[4], double vec_out[4]);
 
 #endif
